Move merge logic to separate file
