+++ packages/astro/src/core/config/config.ts
@@ -5,9 +5,7 @@ import fs from 'fs';
 import * as colors from 'kleur/colors';
 import path from 'path';
 import { fileURLToPath, pathToFileURL } from 'url';
-import { mergeConfig as mergeViteConfig } from 'vite';
 import { AstroError, AstroErrorData } from '../errors/index.js';
-import { arraify, isObject, isURL } from '../util.js';
 import { createRelativeSchema } from './schema.js';
 import { loadConfigWithVite } from './vite-load.js';
 
@@ -294,55 +292,4 @@ export function createDefaultDevConfig(
 	root: string = process.cwd()
 ) {
 	return resolveConfig(userConfig, root, undefined, 'dev');
-}
-
-function mergeConfigRecursively(
-	defaults: Record<string, any>,
-	overrides: Record<string, any>,
-	rootPath: string
-) {
-	const merged: Record<string, any> = { ...defaults };
-	for (const key in overrides) {
-		const value = overrides[key];
-		if (value == null) {
-			continue;
-		}
-
-		const existing = merged[key];
-
-		if (existing == null) {
-			merged[key] = value;
-			continue;
-		}
-
-		// fields that require special handling:
-		if (key === 'vite' && rootPath === '') {
-			merged[key] = mergeViteConfig(existing, value);
-			continue;
-		}
-
-		if (Array.isArray(existing) || Array.isArray(value)) {
-			merged[key] = [...arraify(existing ?? []), ...arraify(value ?? [])];
-			continue;
-		}
-		if (isURL(existing) && isURL(value)) {
-			merged[key] = value;
-			continue;
-		}
-		if (isObject(existing) && isObject(value)) {
-			merged[key] = mergeConfigRecursively(existing, value, rootPath ? `${rootPath}.${key}` : key);
-			continue;
-		}
-
-		merged[key] = value;
-	}
-	return merged;
-}
-
-export function mergeConfig(
-	defaults: Record<string, any>,
-	overrides: Record<string, any>,
-	isRoot = true
-): Record<string, any> {
-	return mergeConfigRecursively(defaults, overrides, isRoot ? '' : '.');
 }
